fix: Correct HTML option tag syntax in orders.php and improve cart scripts for better error handling

+++ admin/script/delete-cart.php
@@ -2,12 +2,15 @@
 
     include "config.php";
     $DELETE_ID = mysqli_real_escape_string($conn, $_POST['delete_id']);
-
-    $query = "DELETE FROM cart WHERE id = {$DELETE_ID}";
-    $result = mysqli_query($conn, $query);
-    if($result){
-        echo "Product deleted from cart";
-        die();
+    if($_SERVER['REQUEST_METHOD'] === 'POST'){
+        $query = "DELETE FROM cart WHERE id = {$DELETE_ID}";
+        $result = mysqli_query($conn, $query);
+        if($result){
+            echo "Product deleted from cart";
+            die();
+        }
+    }else{
+        header("Location: {$host_name}/admin/not-found.php");
     }
 
 ?>
